fix: cython 3 compatibility

+++ bencoder.pyx
@@ -149,7 +149,7 @@ cdef encode_dict(x, list r):
 encode_func = {
     int: encode_int,
     bool: encode_int,    
-    long: encode_long,
+    int: encode_long,
     bytes: encode_bytes,
     str: encode_string,
     list: encode_list,
